requirements.txt
Flask
pandas
fpdf
Pillow
Flask-Mail
reportlab
requests
gunicorn
pymongo 
dnspython
python-dotenv
OTPLessAuthSDK
apscheduler
PyPDF2
 
gspread
google-auth
google-api-python-client

plotly
boto3